+++ CHANGELOG.md
@@ -6,8 +6,8 @@ description: |-
 ---
 
 # VKCS Provider's changelog
-#### v0.13.3 (unreleased)
-- Fix terraform version in the README.md
+#### v0.14.0 (unreleased)
+- Add vkcs_iam_s3_account resource
 
 #### v0.13.2
 - Add details to the vkcs_compute_instance creation error